+++ src/Text/Ascii/Internal.hs
@@ -26,11 +26,27 @@ import Data.CaseInsensitive (FoldCase (foldCase))
 import Data.Char (chr, isAscii)
 import Data.Coerce (coerce)
 import Data.Hashable (Hashable)
+import qualified Data.List.NonEmpty as NE
 import Data.Word (Word8)
 import GHC.Exts (IsList (Item, fromList, fromListN, toList))
 import Numeric (showHex)
 import Optics.AffineTraversal (An_AffineTraversal, atraversal)
 import Optics.At.Core (Index, IxValue, Ixed (IxKind, ix))
+import Text.Megaparsec.Stream
+  ( Stream
+      ( Token,
+        Tokens,
+        chunkLength,
+        chunkToTokens,
+        take1_,
+        takeN_,
+        takeWhile_,
+        tokenToChunk,
+        tokensToChunk
+      ),
+    TraversableStream (reachOffset),
+    VisualStream (showTokens),
+  )
 import Type.Reflection (Typeable)
 
 -- | Represents valid ASCII characters, which are bytes from @0x00@ to @0x7f@.
@@ -150,6 +166,7 @@ instance Ixed AsciiText where
           Nothing -> AsciiText at
           Just (_, end') -> AsciiText (lead <> BS.singleton ac <> end')
 
+-- | @since 1.0.1
 instance FoldCase AsciiText where
   {-# INLINEABLE foldCase #-}
   foldCase (AsciiText bs) = AsciiText . BS.map go $ bs
@@ -158,6 +175,38 @@ instance FoldCase AsciiText where
       go w8
         | 65 <= w8 && w8 <= 90 = w8 + 32
         | otherwise = w8
+
+-- | @since 1.0.1
+instance Stream AsciiText where
+  type Token AsciiText = AsciiChar
+  type Tokens AsciiText = AsciiText
+  {-# INLINEABLE tokenToChunk #-}
+  tokenToChunk _ = coerce BS.singleton
+  {-# INLINEABLE tokensToChunk #-}
+  tokensToChunk _ = fromList
+  {-# INLINEABLE chunkToTokens #-}
+  chunkToTokens _ = toList
+  {-# INLINEABLE chunkLength #-}
+  chunkLength _ = coerce BS.length
+  {-# INLINEABLE take1_ #-}
+  take1_ = coerce BS.uncons
+  {-# INLINEABLE takeN_ #-}
+  takeN_ n at@(AsciiText bs)
+    | n <= 0 = Just (coerce BS.empty, at)
+    | BS.length bs == 0 = Nothing
+    | otherwise = Just . coerce . BS.splitAt n $ bs
+  {-# INLINEABLE takeWhile_ #-}
+  takeWhile_ = coerce BS.span
+
+-- | @since 1.0.1
+instance VisualStream AsciiText where
+  {-# INLINEABLE showTokens #-}
+  showTokens _ = fmap (chr . fromIntegral) . coerce @_ @[Word8] . NE.toList
+
+-- | @since 1.0.1
+instance TraversableStream AsciiText where
+  {-# INLINEABLE reachOffset #-}
+  reachOffset o ps = coerce (reachOffset o ps)
 
 -- Helpers