+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Expose `Text.Ascii.Internal` and `Text.Ascii.QQ`.
 * Add `Ixed` instance (and supporting type instances) for `AsciiText`.
+* Add `Stream`, `VisualStream` and `TraversableStream` instances (and supporting
+  type instances) for `AsciiText`.
+* Drop Parsec in favour of Megaparsec.
 * Add `FoldCase` instances for `AsciiChar` and `AsciiText`.
 * Implement `lines`, `unlines`, `words`, `unwords`, `replicate`, `chunksOf`, 
   `index`, `zipWith` for `AsciiText`.